pyproject.toml
[project]
name = "prompt-optimizer"
version = "0.1.0"
description = "LangGraph 기반 HITL 프롬프트 최적화 도구"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = []
